♻️ move parseOptions to options.go
and rename args_test.go to options_test.go
