+++ options.go
@@ -1,5 +1,10 @@
 package main
 
+import (
+	"fmt"
+	"strings"
+)
+
 type OptionType struct {
 	ShortName string
 	LongName  string
@@ -41,3 +46,36 @@ type Option struct {
 }
 
 type Options []*Option
+
+func parseOptions(args []string) (Options, []string, error) {
+	options := Options{}
+	commandArgs := []string{}
+	inOptions := true
+	var waitingOption *Option
+	for _, arg := range args {
+		if waitingOption != nil {
+			waitingOption.Value = arg
+			options = append(options, waitingOption)
+			waitingOption = nil
+			continue
+		}
+		if inOptions && strings.HasPrefix(arg, "-") {
+			optionType, ok := optionKeyMap[arg]
+			if !ok {
+				return nil, nil, fmt.Errorf("Unknown option: %s", arg)
+			}
+			if optionType.HasValue {
+				waitingOption = &Option{Type: optionType}
+			} else {
+				options = append(options, &Option{Type: optionType})
+			}
+		} else {
+			inOptions = false
+			commandArgs = append(commandArgs, arg)
+		}
+	}
+	if waitingOption != nil {
+		return nil, nil, fmt.Errorf("no value given for option %s", waitingOption.Type.LongName)
+	}
+	return options, commandArgs, nil
+}
